mv format_duration to dagster-shared
